requirements.txt
git+https://github.com/Rapptz/discord.py
praw == 7.5.0
PyNaCl==1.4.0
pycparser==2.20
discord.py
chardet==4.0.0
python-dotenv==0.19.0
toml==0.10.2
